apps/api/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
SQLAlchemy>=2.0
httpx>=0.27
python-dotenv>=1.0
websockets>=12.0
claude-code-sdk>=0.0.20
docker>=7.0
cryptography>=42.0
openai>=1.40
unidiff>=0.7
aiohttp>=3.9
rich>=13.0
python-multipart>=0.0.6
psycopg2-binary>=2.9
alembic>=1.13
python-jose[cryptography]>=3.3.0
stripe>=9.0.0
pydantic-core>=2.20
asyncpg>=0.29